requirements.txt
# CrewAI core and tools
crewai>=0.30.2
crewai-tools>=0.14.0

# Selenium for dynamic web scraping (used by SeleniumScrapingTool)
selenium>=4.18.1

# BeautifulSoup for HTML parsing
beautifulsoup4>=4.12.3

# Requests for HTTP requests
requests>=2.31.0

# Python-dotenv for environment variable management
python-dotenv>=1.0.0

# Logging (standard library, no need to install)
# pathlib (standard library, no need to install)

# For compatibility with CrewAI and Python 3.11+
pydantic>=2.5.0,<3.0.0

# If you use Chrome with Selenium, you may also want:
webdriver-manager>=4.0.0

# Optional: tqdm for progress bars (if you add them)
tqdm>=4.66.0
